Using basic formulas
Used some basic formulas to find the total (by addition of the each month sales), the average sales through the four months (by addition and then division by the number of total months) and finally percentage of difference between two months (by subtraction and then division of the month I want to compare with). In the process I also found that I can find errors in spreadsheet which I need to deal with to ensure the formulas are working properly.